Point 13: Rename
- Renamed global_manager.py to capice_manager.py.
