+++ src/main/python/cli/args_handler_predict.py
@@ -4,7 +4,7 @@ import xgboost as xgb
 
 from src.main_predict import Predict
 from src.main.python.__version__ import __version__
-from src.main.python.core.global_manager import CapiceManager
+from src.main.python.core.capice_manager import CapiceManager
 from src.main.python.cli.args_handler_parent import ArgsHandlerParent
