+++ src/main/python/core/capice_manager.py
