+++ src/main/python/core/logger.py
@@ -21,7 +21,7 @@
 
 import sys
 import logging
-from src.main.python.core.global_manager import CapiceManager
+from src.main.python.core.capice_manager import CapiceManager
 from src.main.python.utilities.custom_log_filter import CustomLoggingFilter
