+++ src/main/python/utilities/preprocessor.py
@@ -2,7 +2,7 @@ import numpy as np
 import pandas as pd
 from src.main.python.core.logger import Logger
 from src.main.python.utilities.sections import Column
-from src.main.python.core.global_manager import CapiceManager
+from src.main.python.core.capice_manager import CapiceManager
 
 
 class PreProcessor:
