+++ src/main_capice.py
@@ -2,7 +2,7 @@ from abc import ABC, abstractmethod
 from src.main.python.core.logger import Logger
 from src.main.python.core.exporter import Exporter
 from src.main.python.utilities.sections import Column
-from src.main.python.core.global_manager import CapiceManager
+from src.main.python.core.capice_manager import CapiceManager
 from src.main.python.utilities.processor import Processor
 from src.main.python.utilities.input_parser import InputParser
 from src.main.python.utilities.capice_imputing import CapiceImputing
