+++ src/test/python/test_templates.py
@@ -2,7 +2,7 @@ import os
 import pickle
 from src.main_predict import Predict
 from src.main.python.core.logger import Logger
-from src.main.python.core.global_manager import CapiceManager
+from src.main.python.core.capice_manager import CapiceManager
 from src.main.python.utilities.utilities import get_project_root_dir
